Added a user input for digit drawing
Using a pygame UI, obtain a drawing of a digit
from the user and use the model with pretrained
weights to predict what digit was drawn.

A TypeError is currently being raised by main.py,
states: "Expected state_dict to be dict-like,
got <class 'method'>". Working on debugging this
ASAP.